project-code/requirements.txt
connexion
pyyaml
